sports_project/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
pandas>=2.1.0
numpy>=1.26.0
Pillow>=10.0.0
